pyproject.toml
[tool.poetry]
name = "smartpharmacy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<=3.12"
django = "^4.2.7"
django-cors-headers = "^4.3.0"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
